+++ src/main/java/com/codetaylor/mc/pyrotech/modules/tech/basic/tile/spi/TileAnvilBase.java
@@ -146,7 +146,9 @@ public abstract class TileAnvilBase
   // - Subclass
   // ---------------------------------------------------------------------------
 
-  public abstract int getBloomAnvilDamagePerHit();
+  public abstract int getBloomAnvilExtraDamagePerHit();
+
+  public abstract double getBloomAnvilExtraDamageChance();
 
   protected abstract int getHitsPerDamage();